Create PGS_Processing

+++ src/main/java/micycle/pgs/utility/RandomPolygon.java
@@ -22,15 +22,15 @@ public class RandomPolygon {
 	 * @param
 	 * @return
 	 */
-	public static List<PVector> generateRandomConvexPolygon(int n, float xMax, float yMax) {
+	public static List<PVector> generateRandomConvexPolygon(int n, double xMax, double yMax) {
 		// Generate two lists of random X and Y coordinates
 
-		List<Float> xPool = new ArrayList<>(n);
-		List<Float> yPool = new ArrayList<>(n);
+		List<Double> xPool = new ArrayList<>(n);
+		List<Double> yPool = new ArrayList<>(n);
 
 		for (int i = 0; i < n; i++) {
-			xPool.add(RAND.nextFloat() * yMax);
-			yPool.add(RAND.nextFloat() * yMax);
+			xPool.add(RAND.nextDouble() * yMax);
+			yPool.add(RAND.nextDouble() * yMax);
 		}
 
 		// Sort them
@@ -38,19 +38,19 @@ public class RandomPolygon {
 		Collections.sort(yPool);
 
 		// Isolate the extreme points
-		float minX = xPool.get(0);
-		float maxX = xPool.get(n - 1);
-		float minY = yPool.get(0);
-		float maxY = yPool.get(n - 1);
+		double minX = xPool.get(0);
+		double maxX = xPool.get(n - 1);
+		double minY = yPool.get(0);
+		double maxY = yPool.get(n - 1);
 
 		// Divide the interior points into two chains & Extract the vector components
-		List<Float> xVec = new ArrayList<>(n);
-		List<Float> yVec = new ArrayList<>(n);
+		List<Double> xVec = new ArrayList<>(n);
+		List<Double> yVec = new ArrayList<>(n);
 
-		float lastTop = minX, lastBot = minX;
+		double lastTop = minX, lastBot = minX;
 
 		for (int i = 1; i < n - 1; i++) {
-			float x = xPool.get(i);
+			double x = xPool.get(i);
 
 			if (RAND.nextBoolean()) {
 				xVec.add(x - lastTop);
@@ -64,10 +64,10 @@ public class RandomPolygon {
 		xVec.add(maxX - lastTop);
 		xVec.add(lastBot - maxX);
 
-		float lastLeft = minY, lastRight = minY;
+		double lastLeft = minY, lastRight = minY;
 
 		for (int i = 1; i < n - 1; i++) {
-			float y = yPool.get(i);
+			double y = yPool.get(i);
 
 			if (RAND.nextBoolean()) {
 				yVec.add(y - lastLeft);
@@ -88,20 +88,20 @@ public class RandomPolygon {
 		List<PVector> vec = new ArrayList<>(n);
 
 		for (int i = 0; i < n; i++) {
-			vec.add(new PVector(xVec.get(i), yVec.get(i)));
+			vec.add(new PVector(xVec.get(i).floatValue(), yVec.get(i).floatValue()));
 		}
 
 		// Sort the vectors by angle
 		Collections.sort(vec, Comparator.comparingDouble(v -> Math.atan2(v.y, v.x)));
 
 		// Lay them end-to-end
-		float x = 0, y = 0;
-		float minPolygonX = 0;
-		float minPolygonY = 0;
+		double x = 0, y = 0;
+		double minPolygonX = 0;
+		double minPolygonY = 0;
 		List<PVector> points = new ArrayList<>(n);
 
 		for (int i = 0; i < n; i++) {
-			points.add(new PVector(x, y));
+			points.add(new PVector((float) x, (float) y));
 
 			x += vec.get(i).x;
 			y += vec.get(i).y;
@@ -111,12 +111,12 @@ public class RandomPolygon {
 		}
 
 		// Move the polygon to the original min and max coordinates
-		float xShift = minX - minPolygonX;
-		float yShift = minY - minPolygonY;
+		double xShift = minX - minPolygonX;
+		double yShift = minY - minPolygonY;
 
 		for (int i = 0; i < n; i++) {
 			PVector p = points.get(i);
-			points.set(i, new PVector(p.x + xShift, p.y + yShift));
+			points.set(i, new PVector((float) (p.x + xShift), (float) (p.y + yShift)));
 		}
 
 		return points;
